fix(lint): resolve all PHPStan errors for PHP 8.4 compatibility
Fixed 15 PHPStan static analysis errors across cartridge implementation:

**Type Safety Fixes:**
1. **Division Return Types**: Cast division operations to int to satisfy strict typing
   - CartridgeHeader::getRamBankCount(): Added (int) cast
   - Mbc1/Mbc3/Mbc5 constructors: Cast romBankCount calculation to int
   - Prevents float|int union type issues

2. **SaveManager Type Assertions**:
   - Added is_array() check for json_decode($data['rtc'])
   - Added PHPDoc @var annotation for $rtcState type hint
   - Handle preg_replace() returning string|null with null coalescing

3. **Test Method Annotations**:
   - Added @return array<int, int> to all createRom() helper methods
   - Added @param and @return annotations to CartridgeHeaderTest::createRomWithHeader()
   - Ensures PHPStan understands array value types

4. **SaveManagerTest glob() Handling**:
   - Added check for glob() returning false (rare error case)
   - Prevents foreach on non-iterable type error

**PHPStan Results:**
- ✅ All 15 errors resolved
- ✅ Zero warnings
- ✅ Strict type checking enforced
- ✅ Full compatibility with PHP 8.4+

**Test Results:**
- ✅ All 67 tests passing
- ✅ 12,485 assertions verified
- ✅ No behavioral changes

**Files Modified:**
- CartridgeHeader.php: getRamBankCount() return type
- Mbc1/3/5.php: romBankCount calculation
- SaveManager.php: RTC state types, preg_replace null handling
- Test files: Added PHPDoc type annotations

Verified with: php -d memory_limit=512M vendor/bin/phpstan analyse

+++ src/Cartridge/CartridgeHeader.php
@@ -246,7 +246,7 @@ final readonly class CartridgeHeader
             return 0;
         }
 
-        return $ramSize / (8 * 1024);
+        return (int)($ramSize / (8 * 1024));
     }
 
     /**
